issue(#2): ahoi-serial-lib
- Major refactoring

- Add possibility to send command

- Reception is now in sync

- Adapt reception to the ahoi firmware
