+++ config.cmake
@@ -9,6 +9,6 @@ set(TEST_LIST "" CACHE STRING "")
 # App config
 option(WITH_SENDER "Build sender" ON)
 option(WITH_RECEIVER "Build receiver" ON)
-option(ENABLE_TESTS "Enable testing" ON)
+option(ENABLE_TESTS "Enable testing" OFF)
 
 set(AHOI_LIB_NAME "ahoi-serial-lib" CACHE STRING "Library name")
