--- include/pub/ahoi_serial/common_defs.h
@@ -1,12 +1,0 @@
-#ifndef AHOI_SERIAL_COMMON_DEFS_H
-#define AHOI_SERIAL_COMMON_DEFS_H
-
-#define KEY_SIZE 16
-#define NONCE_SIZE 16
-#define TAG_SIZE 16
-#define HEADER_SIZE 6
-#define MAX_PAYLOAD_SIZE 256
-#define MAX_SECURE_PAYLOAD_SIZE (256 - TAG_SIZE)
-#define MAX_PACKET_SIZE (HEADER_SIZE + MAX_PAYLOAD_SIZE)
-
-#endif
