+++ include/pub/ahoilib.h
@@ -1,0 +1,9 @@
+#ifndef AHOILIB_H
+#define AHOILIB_H
+
+#include "ahoi_serial/ahoi_defs.h"
+#include "ahoi_serial/core.h"
+#include "ahoi_serial/com.h"
+#include "ahoi_serial/cmd.h"
+
+#endif //AHOILIB_H
