--- src/ahoi_commons/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_library(${COMMONS_LIB_NAME} OBJECT
-        "core.c"
-        "security.c")
-
-target_include_directories(${COMMONS_LIB_NAME} PRIVATE "${PROJECT_SOURCE_DIR}/include/pub/ahoi_serial")
-target_include_directories(${COMMONS_LIB_NAME} PRIVATE "${PROJECT_SOURCE_DIR}/include/priv")
-
-target_link_libraries(${COMMONS_LIB_NAME} PRIVATE ${ASCON_LIB_NAME})
