--- src/ahoi_receiver/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_library(${RECEIVER_LIB_NAME} OBJECT "receiver.c")
-
-target_include_directories(${RECEIVER_LIB_NAME} PRIVATE "${PROJECT_SOURCE_DIR}/include/pub/ahoi_serial")
-target_include_directories(${RECEIVER_LIB_NAME} PRIVATE "${PROJECT_SOURCE_DIR}/include/priv")
-
-target_sources(${RECEIVER_LIB_NAME} PRIVATE $<TARGET_OBJECTS:${COMMONS_LIB_NAME}>)
